+++ basic_autonomy/include/basic_autonomy/basic_autonomy.h
@@ -1,6 +1,6 @@
 #pragma once
 /*
- * Copyright (C) 2019-2020 LEIDOS.
+ * Copyright (C) 2021 LEIDOS.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
